Frame prediction

+++ backend/app/api/routes.py
@@ -2,17 +2,13 @@ from flask import Flask, flash, request, url_for, redirect
 from flask import current_app as app
 from werkzeug.utils import secure_filename
 from app.services.extract_frames import extract_frames
-from app.services.process_frames import preprocess_frames, get_string_from_frames
+from app.services.is_chosen import is_chosen
+from app.services.process_frames import preprocess_frames, get_string_from_frames, get_last_number
 from app.model.predict import predict_frets
 import os
 def allowed_file(filename):
     return '.' in filename and \
            filename.rsplit('.', 1)[1].lower() in app.config["ALLOWED_EXTENSIONS"]
-
-# Fetch the last number of the frames
-def get_last_number(frame: str) -> int:
-    print(frame)
-    return int(frame.split('_')[-1].split('.')[0])
 
 def register_routes(app):
     @app.route("/api/upload", methods = ["POST"])
@@ -21,22 +17,27 @@ def register_routes(app):
             return {"message": "Missing 'video' in request.files", "status": 400}
 
         video = request.files['video']
-
+        new_line_per_second = int(request.form.get("new_line"))
         if not allowed_file(video.filename):
             return {"message":"File format not allowed", "status": 404}
 
         
         filename = secure_filename(video.filename)
         video.save(os.path.join(app.config['UPLOAD_FOLDER'], filename))
-        extract_frames(os.path.join(app.config['UPLOAD_FOLDER'], filename))
+        extract_frames(os.path.join(app.config['UPLOAD_FOLDER'], filename), new_line_per_second)
+        if not os.path.exists(os.path.join(app.config['UPLOAD_FOLDER'], filename)):
+            return {"message": "Could not extract frames", "status": 400}
         os.remove(os.path.join(app.config['UPLOAD_FOLDER'], filename))  
         return {"success":True, "status": 200}
 
     @app.route("/api/get_frames", methods = ["GET"])
     def get_frames():
-        frames = sorted(os.listdir(app.config['FRAMES_FOLDER']), key = lambda x: get_last_number(x))
+        # filter out the chosen_frames.json file
+        frames = [frame for frame in os.listdir(app.config['FRAMES_FOLDER']) if frame.endswith('.jpg')]
+        frames = sorted(frames, key = lambda x: get_last_number(x))
+
         # return file names
-        return [f"/static/frames/{frame}" for frame in frames]
+        return [(f"/static/frames/{frame}", is_chosen(frame)) for frame in frames]
     
     # Get the confirmed frames and send in as a list of numbers
     @app.route("/api/confirmed_frames", methods = ["POST"])
@@ -51,5 +52,7 @@ def register_routes(app):
         strings_capture_result = get_string_from_frames()
         if not strings_capture_result["success"]:
             return strings_capture_result
-        predict_frets()
+        predict_frets_result = predict_frets()
+        if not predict_frets_result["success"]:
+            return {"message": "Could not predict frets", "success": False, "status": 400}
         return {"success":True, "status": 200}
